+++ src/components/stories/LoginButton.stories.js
@@ -2,8 +2,8 @@ import React from 'react';
 import  { storiesOf } from '@storybook/react';
 import { boolean } from '@storybook/addon-knobs';
 
-import LoginButton from '../../../components/LoginButton/LoginButton';
-import translateMock from '../../../utils/translate';
+import LoginButton from '../LoginButton/LoginButton';
+import translateMock from '../../utils/translate';
 
 import logIn from '../../../images/log-in.svg';
