+++ src/components/stories/styles.scss
