+++ src/modules/Picture/stories/picture1.png
