+++ src/modules/Table/stories/Table.stories.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import { boolean, object } from '@storybook/addon-knobs';
 import { action }  from '@storybook/addon-actions';
 
-import Table from '../../../modules/Table/Table';
+import Table from '../Table';
 
 const columns = [
   'foo',
