+++ src/modules/Visualizer/stories/layersTree.js
@@ -1,4 +1,4 @@
-import { TYPE_SINGLE, TYPE_MANY, TYPE_RANGE, TYPE_BOOL } from '../../../modules/Forms/Filters/index';
+import { TYPE_SINGLE, TYPE_MANY, TYPE_RANGE, TYPE_BOOL } from '../../Forms/Filters/index';
 
 export const storyLayersTreeConfig = [
   {
