+++ src/stories/index.js
@@ -4,4 +4,4 @@ import '@blueprintjs/core/lib/css/blueprint.css';
 import '@blueprintjs/datetime/lib/css/blueprint-datetime.css';
 
 import './summary';
-import './modules/Map';
+import '../modules/Map/stories/';
